refactor

+++ common/models/responsemodel.go
@@ -1,21 +1,11 @@
 package models
 
-// swagger:model ErrorResponse
 type ErrorResponse struct {
-	// Status of the error
-	// in: int64
-	Status int64 `json:"status"`
-	// Message of the error
-	// in: string
+	Status  int    `json:"status"`
 	Message string `json:"message"`
 }
 
-// swagger:model SuccessResponse
 type SuccessResponse struct {
-	// Status of the error
-	// in: int64
-	Status int64 `json:"status"`
-	// Message of the error
-	// in: string
+	Status  int    `json:"status"`
 	Message string `json:"message"`
 }
